// src/gql/graphql.ts
/* eslint-disable */
import { TypedDocumentNode as DocumentNode } from '@graphql-typed-document-node/core'
import gql from 'graphql-tag'
import * as Urql from 'urql'
export type Maybe<T> = T | null
export type InputMaybe<T> = Maybe<T>
export type Exact<T extends { [key: string]: unknown }> = {
    [K in keyof T]: T[K]
}
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
    [SubKey in K]?: Maybe<T[SubKey]>
}
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
    [SubKey in K]: Maybe<T[SubKey]>
}
export type Omit<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
    ID: string
    String: string
    Boolean: boolean
    Int: number
    Float: number
}

export type FieldError = {
    __typename?: 'FieldError'
    field: Scalars['String']
    message: Scalars['String']
}

export type Mutation = {
    __typename?: 'Mutation'
    createPost: Post
    deletePost: Scalars['Boolean']
    login: UserResponse
    register: UserResponse
    updatePost?: Maybe<Post>
}

export type MutationCreatePostArgs = {
    title: Scalars['String']
}

export type MutationDeletePostArgs = {
    id: Scalars['Float']
}

export type MutationLoginArgs = {
    options: UsernamePasswordInput
}

export type MutationRegisterArgs = {
    options: UsernamePasswordInput
}

export type MutationUpdatePostArgs = {
    id: Scalars['Int']
    title?: InputMaybe<Scalars['String']>
}

export type Post = {
    __typename?: 'Post'
    createdAt: Scalars['String']
    id: Scalars['Int']
    title: Scalars['String']
    updatedAt: Scalars['String']
}

export type Query = {
    __typename?: 'Query'
    currentUser?: Maybe<User>
    hello: Scalars['String']
    post?: Maybe<Post>
    posts: Array<Post>
}

export type QueryPostArgs = {
    id: Scalars['Int']
}

export type User = {
    __typename?: 'User'
    createdAt: Scalars['String']
    id: Scalars['Int']
    updatedAt: Scalars['String']
    username: Scalars['String']
}

export type UserResponse = {
    __typename?: 'UserResponse'
    errors?: Maybe<Array<FieldError>>
    user?: Maybe<User>
}

export type UsernamePasswordInput = {
    password: Scalars['String']
    username: Scalars['String']
}

export type RegisterMutationVariables = Exact<{
    username: Scalars['String']
    password: Scalars['String']
}>

export type RegisterMutation = {
    __typename?: 'Mutation'
    register: {
        __typename?: 'UserResponse'
        user?: { __typename?: 'User'; username: string } | null
        errors?: Array<{
            __typename?: 'FieldError'
            field: string
            message: string
        }> | null
    }
}

export const RegisterDocument = gql`
    mutation Register($username: String!, $password: String!) {
        register(options: { username: $username, password: $password }) {
            user {
                username
            }
            errors {
                field
                message
            }
        }
    }
`

export function useRegisterMutation() {
    return Urql.useMutation<RegisterMutation, RegisterMutationVariables>(
        RegisterDocument
    )
}